+++ webknossos/webknossos/dataset/dataset.py
@@ -1299,6 +1299,39 @@ class Dataset:
         else:
             dtype_per_channel = np.dtype("uint" + str(DEFAULT_BIT_DEPTH))
 
+        # assert that the dtype_per_channel is supported by webknossos
+        if category == COLOR_CATEGORY:
+            color_dtypes = (
+                "uint8",
+                "uint16",
+                "uint32",
+                "int8",
+                "int16",
+                "int32",
+                "float32",
+            )
+            if dtype_per_channel.name not in color_dtypes:
+                raise ValueError(
+                    f"Cannot add color layer with dtype {dtype_per_channel.name}. "
+                    f"Supported dtypes are: {', '.join(color_dtypes)}.",
+                )
+        else:
+            segmentation_dtypes = (
+                "uint8",
+                "uint16",
+                "uint32",
+                "uint64",
+                "int8",
+                "int16",
+                "int32",
+                "int64",
+            )
+            if dtype_per_channel.name not in segmentation_dtypes:
+                raise ValueError(
+                    f"Cannot add segmentation layer with dtype {dtype_per_channel.name}. "
+                    f"Supported dtypes are: {', '.join(segmentation_dtypes)}.",
+                )
+
         if layer_name in self.layers.keys():
             raise IndexError(
                 f"Adding layer {layer_name} failed. There is already a layer with this name"